Parte Julian

+++ src/main/java/com/Gropo06/turnos_medicos/controller/EspecialidadController.java
@@ -1,7 +1,11 @@
 package com.Gropo06.turnos_medicos.controller;
 
+import com.Gropo06.turnos_medicos.exceptions.CustomException;
 import com.Gropo06.turnos_medicos.model.Especialidad;
+import com.Gropo06.turnos_medicos.repository.AgendaRepository;
 import com.Gropo06.turnos_medicos.repository.EspecialidadRepository;
+
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
@@ -13,6 +17,9 @@ import java.util.List;
 public class EspecialidadController {
 
 	private final EspecialidadRepository especialidadRepo;
+	
+	@Autowired
+	private AgendaRepository agendaRepo;
 
 	public EspecialidadController(EspecialidadRepository especialidadRepo) {
 		this.especialidadRepo = especialidadRepo;
@@ -34,6 +41,9 @@ public class EspecialidadController {
 
 	@PostMapping("/save")
 	public String guardarEspecialidad(@ModelAttribute("especialidadForm") Especialidad esp) {
+		if (especialidadRepo.existsByNombre(esp.getNombre())) {
+		    throw new CustomException("Nombre Duplicado");
+		}
 		especialidadRepo.save(esp);
 		return "redirect:/empleado/especialidades";
 	}
@@ -43,7 +53,7 @@ public class EspecialidadController {
 	@GetMapping("/edit/{id}")
 	public String editarEspecialidad(@PathVariable("id") Long id, Model model) {
 		Especialidad esp = especialidadRepo.findById(id)
-				.orElseThrow(() -> new IllegalArgumentException("Especialidad inválida Id:" + id));
+				.orElseThrow(() -> new CustomException("Especialidad inválida Id:" + id));
 		model.addAttribute("especialidadForm", esp);
 
 		// Recargamos la lista
@@ -57,6 +67,7 @@ public class EspecialidadController {
 
 	@GetMapping("/delete/{id}")
 	public String eliminarEspecialidad(@PathVariable("id") Long id) {
+		if (agendaRepo.existsByTipoEspecialidad_IdEspecialidad(id)) throw new CustomException("Especialidad ya tiene una Agenda asignada.");
 		especialidadRepo.deleteById(id);
 		return "redirect:/empleado/especialidades";
 	}
